location: Move private declaration away from header
