+++ src/detect-location.c
@@ -38,6 +38,20 @@
 #include <libloc/database.h>
 #include <libloc/network.h>
 
+enum location_flags {
+    LOCATION_FLAG_SRC     = (1 << 0),
+    LOCATION_FLAG_DST     = (1 << 1),
+    LOCATION_FLAG_BOTH    = (1 << 2),
+    LOCATION_FLAG_NEGATED = (1 << 3),
+};
+
+struct DetectLocationData {
+    struct loc_ctx* ctx;
+    struct loc_database* db;
+    char** countries;
+    int flags;
+};
+
 /**
  * \brief This function will free all resources used by the location module
  *
