+++ src/detect-location.h
@@ -26,27 +26,6 @@
 #ifndef __DETECT_LOCATION_H__
 #define __DETECT_LOCATION_H__
 
-#ifdef HAVE_LIBLOC
-
-#include <libloc/libloc.h>
-#include <libloc/database.h>
-
-enum location_flags {
-    LOCATION_FLAG_SRC     = (1 << 0),
-    LOCATION_FLAG_DST     = (1 << 1),
-    LOCATION_FLAG_BOTH    = (1 << 2),
-    LOCATION_FLAG_NEGATED = (1 << 3),
-};
-
-struct DetectLocationData {
-    struct loc_ctx* ctx;
-    struct loc_database* db;
-    char** countries;
-    int flags;
-};
-
-#endif /* HAVE_LIBLOC */
-
 void DetectLocationRegister(void);
 
 #endif /* __DETECT_LOCATION_H__ */
